requirements.txt
biopython==1.83
matplotlib==3.9.0

# Clustal Omega (1.2.4) needs to be installed separately. 
# Instructions: On Ubuntu/Debian-based Linux run `sudo apt-get install clustalo`
